Modify User route and security.yml, fixed some minor bugs involving the header
