+++ templates/admin/user/edit.html.twig
@@ -1,0 +1,13 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Edit User{% endblock %}
+
+{% block body %}
+    <h1>Edit User</h1>
+
+    {{ include('admin/user/_form.html.twig', {'button_label': 'Update'}) }}
+
+    <a href="{{ path('app_admin_user_index') }}">back to list</a>
+
+    {{ include('admin/user/_delete_form.html.twig') }}
+{% endblock %}
