--- templates/user/_form.html.twig
@@ -1,15 +1,0 @@
-<div class="card">
-    <div class="card-body">
-        {% if user %}
-            {{ form_start(form) }}
-                {{ form_widget(form) }}
-            {{ form_end(form) }}
-        {% else %}
-            {{ form_start(form) }}
-                {{ form_widget(form) }}
-            {{ form_end(form) }}
-        {% endif %}
-    </div>
-</div>
-
- <button class="btn">{{ button_label|default('Save') }}</button>
